requirements.txt
absl-py==1.3.0
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
beautifulsoup4==4.11.1
black==22.3.0
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
cryptography==36.0.0
cycler==0.11.0
Cython==0.29.32
cytoolz==0.11.2
dask==2022.4.2
docker-pycreds==0.4.0
eagerpy==0.30.0
efficientnet-pytorch==0.6.3
fairscale==0.4.12
filelock==3.6.0
fonttools==4.33.3
foolbox==3.3.3
fsspec==2022.3.0
future==0.18.2
fvcore==0.1.5.post20220512
gdown==4.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
h5py==3.6.0
hydra-core==1.2.0
idna==3.3
imagecodecs==2022.2.22
imagecorruptions==1.1.2
imageio==2.18.0
iopath==0.1.9
joblib==1.1.0
kiwisolver==1.4.2
kornia==0.6.3
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
munch==2.5.0
munkres==1.1.4
mypy-extensions==0.4.3
networkx==2.8
numpy==1.23.4
oauthlib==3.2.2
omegaconf==2.2.3
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.2
partd==1.2.0
pathspec==0.10.1
pathtools==0.1.2
Pillow==9.1.0
pip==22.3
platformdirs==2.5.2
portalocker==2.6.0
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.19.6
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pycparser==2.21
pydot==1.4.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.12.1
segmentation-models-pytorch==0.2.1
sentry-sdk==1.5.10
setproctitle==1.2.3
setuptools==62.1.0
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2022.4.28
timm==0.4.12
tomli==2.0.1
toolz==0.11.2
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchmetrics==0.10.0
torchvision==0.13.1+cu116
tqdm==4.64.0
typing_extensions==4.2.0
unicodedata2==14.0.0
urllib3==1.26.9
wandb==0.13.4
Werkzeug==2.2.2
wheel==0.37.1
yacs==0.1.8